code/requirements/base.txt
Django==1.7.11
pytz==2016.4
pylibmc==1.5.1
psycopg2==2.6.1
Pillow==3.2.0
dj-database-url==0.4.1
raven==5.32.0

# Storage
django-storages==1.4.1
boto==2.41.0
blanc-contentfiles==0.2.3

# Assets
blanc-basic-assets==0.3.2

# Admin theme
blanc-admin-theme==1.0.1

# Thumbnails
easy-thumbnails==2.3
six==1.10.0

# Pages
blanc-pages==0.6.6
django-mptt==0.7.4
django-mptt-admin==0.3.7
simplejson==3.8.2

# Blanc page blocks
blanc-pages-redactor-block==0.6.2
blanc-pages-image-block==0.2.3
blanc-pages-form-block==0.3.4

# Redactor
django-redactorjs-staticfiles==9.2.6

# News
blanc-basic-news==0.3.2

# Sortable models
django-admin-sortable==1.8.3

# Advanced forms
django-form-utils==0.3.1

# For jobs/profiles
django-countries==2.0
python-dateutil==2.1

# Pagination on profiles
django-paginationlinks==0.1.1

# Events
blanc-basic-events==0.2.7
vobject==0.9.2

# Mixins
django-braces==1.0.0

# New Relic monitoring
newrelic==2.66.0.49

# Reports
unicodecsv==0.9.4

# API
djangorestframework==3.1.3
django-filter==0.11.0

# Latest tweets
django-latest-tweets==0.4.1
twitter==1.17.1
